fix: svg custom declarations
