+++ .changeset/cool-tables-worry.md
@@ -1,0 +1,5 @@
+---
+'@commencis/starter-react-vite': patch
+---
+
+fix: SVG custom declarations
