+++ src/types/index.ts
@@ -2,5 +2,5 @@ export * from './env';
 export * from './generic';
 export * from './http';
 export * from './localization';
-export * from './svg.d';
+export * from './svg';
 export * from './theme';
